Some proposes fixes to Evolution Preferences
Included here are some suggestions for slightly different layout of
Evolution Preferences to make the preferences window fit properly on a
netbook screen. Included are Add an Account and Edit an Account. (To
follow in another document suggestions for the rest of the preferences
functionality.) NB you will notice that I have included versions of the
screens with dimension suggestions to give a guide to the position of
components to ensure they fit within the screen.